requirements.txt
django == 4.0.2
requests == 2.27.1
django_extensions == 3.1.5
django-import-export ==2.7.1
djangorestframework== 3.13.1
django-filter == 21.1
psycopg2-binary == 2.9.3
drf-yasg2 == 1.19.4